category query now can return ready category seo mix
